elf2rpl: Add explicit narrowing cast to prefix length
This value is a constant 10, anyway - so unless we have a prefix 
approaching a few billion we should be okay.

+++ src/elf2rpl/main.cpp
@@ -551,7 +551,7 @@ renameRplWrap(ElfFile &file)
             std::string(&strtabd[symbol->name]),
             std::string(&strtabd[symbol->name] + rplwrap_prefix.length()));
 #endif //DEBUG
-         symbol->name += rplwrap_prefix.length();
+         symbol->name += (uint32_t)rplwrap_prefix.length();
       }
    }
